requirements.txt
langchain
openai
pypdf
chromadb
tiktoken
langchain-openai
python-dotenv
slack_bolt